refactor: extract SitesSettings panel from SettingsDialog (#672)
First split of the last oversized component. SettingsDialog is a 13-tab dialog
where each tab is a self-contained panel — the natural decomposition is one
component per tab. The privileged-sites tab is the cleanest starting point: its
state (site list + add-form fields) and handlers only ever touch api.sites.*,
so it lifts whole into a self-contained SitesSettings.svelte that owns its
state, loads on mount, and the dialog just mounts it for the "sites" tab.
SettingsDialog 2205 → 2052.

The panel carries its own scoped .field / .hint / input form styles — the
convention every dialog in the app already follows (each scopes its own copy);
the dialog keeps those rules for its other twelve panels.

Behavior note: sites now load when the user opens the Sites tab (the child's
onMount) rather than eagerly on dialog open — equivalent from the user's view
(the list only renders in that panel) and slightly leaner.

Tests (5): components/SitesSettings.test.ts mocks api.sites and pins the on-
mount list render, the empty state, the add-form gating + api.sites.add call,
Login/Logout/Remove → matching IPC with the site id, and the last-login format.

Verified: pnpm lint (0 errors), full suite (2985 passed, +5), electron-forge
build + boot smoke.
